third_party/blink/web_tests/media/controls/accessibility-picture-in-picture-button.html

<!DOCTYPE html>
<html lang="en-US">
<title>Media Controls: picture-in-picture button accessibility tests</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../media-controls.js"></script>
<video controls></video>
<script>
async_test(t => {
  if (!document.pictureInPictureEnabled) {
    t.done();
    return;
  }

  var video = document.querySelector('video');
  video.src = '../content/test.ogv';

  video.oncanplaythrough = t.step_func(_ => {
    var pip_button = pictureInPictureButton(video);

    assert_equals(pip_button.getAttribute('aria-label'),
                  'enter picture-in-picture');
    test_driver.bless('request Picture-in-Picture').then(t.step_func(() => {
      return video.requestPictureInPicture();
    })).then(t.step_func_done(() => {
      assert_equals(pip_button.getAttribute('aria-label'),
                  'exit picture-in-picture');
    })).catch(t.unreached_func('requestPictureInPicture rejected'));
  });
});
</script>
</html>
